Se agregan excepciones personalizadas

+++ src/main/java/com/example/ProyectoExamenPOO/service/EstudianteService.java
@@ -2,9 +2,9 @@ package com.example.ProyectoExamenPOO.service;
 
 import com.example.ProyectoExamenPOO.model.entity.Estudiante;
 import com.example.ProyectoExamenPOO.repository.IEstudianteRepository;
-import com.example.ProyectoExamenPOO.service.IEstudianteService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.example.ProyectoExamenPOO.exception.BadRequestException;
 import org.springframework.stereotype.Service;
+import com.example.ProyectoExamenPOO.exception.ResourceNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -24,26 +24,31 @@ public class EstudianteService implements IEstudianteService {
 
     @Override
     public Estudiante findEstudiante(Long id) {
-        return estudianteRepository.findById(id).orElse(null);
+        return estudianteRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Estudiante no encontrado con ID: " + id));
     }
 
     @Override
     public void saveEstudiante(Estudiante estudiante) {
+        if (estudianteRepository.existsByEmail(estudiante.getEmail())) {
+            throw new BadRequestException("El email " + estudiante.getEmail() + " ya está registrado.");
+        }
         estudianteRepository.save(estudiante);
     }
 
     @Override
     public void deleteEstudiante(Long id) {
+        if (!estudianteRepository.existsById(id)) {
+            throw new ResourceNotFoundException("No se puede eliminar: Estudiante no encontrado con ID: " + id);
+        }
         estudianteRepository.deleteById(id);
     }
 
     @Override
     public Estudiante updateEstudiante(Long id, Estudiante datosNuevos) {
-        Estudiante estudianteExistente = estudianteRepository.findById(id).orElse(null);
+        Estudiante estudianteExistente = estudianteRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Estudiante no encontrado con ID: " + id));
 
-        if (estudianteExistente == null) {
-            return null;
-        }
         if (datosNuevos.getName() != null) {
             estudianteExistente.setName(datosNuevos.getName());
         }
